# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TourVoice API",
    "description": "Hệ thống Thuyết minh Du lịch Tự động Đa ngôn ngữ Quận 4 (FastAPI + PyMongo Async + MongoDB).",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/pois": {
      "get": {
        "tags": [
          "POIs (Points of Interest)"
        ],
        "summary": "List Public Pois",
        "description": "Lists public POIs with language fallback.",
        "operationId": "list_public_pois_api_v1_pois_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Requested language code (vi, en, fr, zh, ja)",
              "default": "vi",
              "title": "Lang"
            },
            "description": "Requested language code (vi, en, fr, zh, ja)"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Public Pois Api V1 Pois Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "POIs (Points of Interest)"
        ],
        "summary": "Create Poi Admin",
        "description": "Creates a new POI (Admin only).",
        "operationId": "create_poi_admin_api_v1_pois_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/POICreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/pois/{poi_id}": {
      "get": {
        "tags": [
          "POIs (Points of Interest)"
        ],
        "summary": "Get Poi Detail",
        "description": "Fetches public POI detail with audio and localization fallback.",
        "operationId": "get_poi_detail_api_v1_pois__poi_id__get",
        "parameters": [
          {
            "name": "poi_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Poi Id"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Requested language code",
              "default": "vi",
              "title": "Lang"
            },
            "description": "Requested language code"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Poi Detail Api V1 Pois  Poi Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "POIs (Points of Interest)"
        ],
        "summary": "Update Poi Admin",
        "description": "Updates POI with optimistic concurrency check.",
        "operationId": "update_poi_admin_api_v1_pois__poi_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "poi_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Poi Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/POIUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "POIs (Points of Interest)"
        ],
        "summary": "Delete Poi Admin",
        "description": "Soft deletes POI.",
        "operationId": "delete_poi_admin_api_v1_pois__poi_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "poi_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Poi Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tours": {
      "get": {
        "tags": [
          "Tours"
        ],
        "summary": "List Tours",
        "description": "Use Case T12: List public tours.",
        "operationId": "list_tours_api_v1_tours_get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tours"
        ],
        "summary": "Create Tour",
        "description": "Use Case C15: Admin creates tour.",
        "operationId": "create_tour_api_v1_tours_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TourCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tours/{tour_id}": {
      "get": {
        "tags": [
          "Tours"
        ],
        "summary": "Get Tour",
        "description": "Use Case T12: View tour detail with ordered stops.",
        "operationId": "get_tour_api_v1_tours__tour_id__get",
        "parameters": [
          {
            "name": "tour_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tour Id"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "vi",
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Tours"
        ],
        "summary": "Update Tour",
        "description": "Use Case C15: Admin updates tour and reorders POIs.",
        "operationId": "update_tour_api_v1_tours__tour_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tour_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tour Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TourUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tours"
        ],
        "summary": "Delete Tour",
        "description": "Use Case C15: Admin soft deletes tour.",
        "operationId": "delete_tour_api_v1_tours__tour_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tour_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tour Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/owner/pois": {
      "get": {
        "tags": [
          "Store Owner Portal (Chủ quán)"
        ],
        "summary": "Get My Pois",
        "description": "Use Case O03: View list of POIs owned by current owner.",
        "operationId": "get_my_pois_api_v1_owner_pois_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GeoLocation": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "Point"
          },
          "coordinates": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Coordinates",
            "description": "[longitude, latitude]"
          }
        },
        "type": "object",
        "required": [
          "coordinates"
        ],
        "title": "GeoLocation"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "POICreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "minLength": 5,
            "title": "Description"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": "sightseeing"
          },
          "address": {
            "type": "string",
            "title": "Address",
            "default": ""
          },
          "location": {
            "$ref": "#/components/schemas/GeoLocation"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Images"
          },
          "trigger_radius": {
            "type": "number",
            "maximum": 500.0,
            "minimum": 5.0,
            "title": "Trigger Radius",
            "default": 30.0
          },
          "audio_priority": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Audio Priority",
            "default": 1
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "source_lang": {
            "type": "string",
            "title": "Source Lang",
            "default": "vi"
          },
          "activation_requested": {
            "type": "boolean",
            "title": "Activation Requested",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "location"
        ],
        "title": "POICreate"
      },
      "POIUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeoLocation"
              },
              {
                "type": "null"
              }
            ]
          },
          "images": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Images"
          },
          "trigger_radius": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 500.0,
                "minimum": 5.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Radius"
          },
          "audio_priority": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Priority"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "expected_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version",
            "description": "Optimistic concurrency control version",
            "default": 1
          }
        },
        "type": "object",
        "title": "POIUpdate"
      },
...
      "TourCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "default": ""
          },
          "localizations": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Localizations"
          },
          "poi_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Poi Ids",
            "description": "Ordered list of POI IDs"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "poi_ids"
        ],
        "title": "TourCreateRequest"
      },
      "TourUpdateRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "localizations": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Localizations"
          },
          "poi_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poi Ids"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "expected_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          }
        },
        "type": "object",
        "title": "TourUpdateRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/admin/auth/login"
          }
        }
      }
    }
  }
}